# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "coco-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
    "/v1/docs/index": {
      "post": {
        "tags": [
          "Docs"
        ],
        "operationId": "index_documents",
        "parameters": [
          {
            "name": "x-coco-org-id",
            "in": "header",
            "description": "Organization identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-coco-project-id",
            "in": "header",
            "description": "Project identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Reindex accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/docs/query": {
      "post": {
        "tags": [
          "Memo"
        ],
        "operationId": "query_documents",
        "parameters": [
          {
            "name": "x-coco-org-id",
            "in": "header",
            "description": "Organization identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-coco-project-id",
            "in": "header",
            "description": "Project identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Query results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponseEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/batch": {
      "post": {
        "tags": [
          "Docs"
        ],
        "operationId": "ingest_batch",
        "parameters": [
          {
            "name": "x-coco-org-id",
            "in": "header",
            "description": "Organization identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-coco-project-id",
            "in": "header",
            "description": "Project identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/memo/query": {
      "post": {
        "tags": [
          "Docs"
        ],
        "operationId": "query_memos",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Memo query results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoQueryResponseEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Chunk": {
        "type": "object",
        "description": "A chunk extracted from a document.",
        "required": [
          "id",
          "doc_id",
          "content",
          "span"
        ],
        "properties": {
          "content": {
            "type": "string",
            "description": "Chunk content."
          },
          "doc_id": {
            "type": "string",
            "description": "Source document identifier."
          },
          "embedding": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "Optional embedding vector.",
            "nullable": true
          },
          "id": {
            "type": "string",
            "description": "Chunk identifier."
          },
          "quality_score": {
            "type": "number",
            "format": "float",
            "description": "Optional quality score for the chunk.",
            "nullable": true
          },
          "span": {
            "$ref": "#/components/schemas/TextSpan"
          },
          "verified": {
            "type": "boolean",
            "description": "Whether the chunk has been verified.",
            "nullable": true
          }
        }
      },
      "CocoErrorKind": {
        "type": "string",
        "description": "High-level error categories for CoCo.\n\nAdding new variants is a breaking change for the public API.",
        "enum": [
          "system",
          "user",
          "network",
          "storage",
          "compute"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Serializable error response used by API layers.",
        "required": [
          "kind",
          "message"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/CocoErrorKind"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "IndexRequest": {
        "type": "object",
        "properties": {
          "indexing_config_id": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "IndexResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "IngestBatchRequest": {
        "type": "object",
        "required": [
          "documents"
        ],
        "properties": {
          "activate": {
            "type": "boolean"
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IngestDocument"
            }
          },
          "indexing_config_id": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "IngestBatchResponse": {
        "type": "object",
        "required": [
          "job_id"
        ],
        "properties": {
          "job_id": {
            "type": "string"
          }
        }
      },
      "IngestChunk": {
        "type": "object",
        "required": [
          "chunk_id",
          "content",
          "embedding",
          "start",
          "end"
        ],
        "properties": {
          "chunk_id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "embedding": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "end": {
            "type": "integer",
            "minimum": 0
          },
          "quality_score": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "start": {
            "type": "integer",
            "minimum": 0
          },
          "verified": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
      "IngestDocument": {
        "type": "object",
        "required": [
          "doc_id",
          "source_ref",
          "chunks"
        ],
        "properties": {
          "chunks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IngestChunk"
            }
          },
          "content_hash": {
            "type": "string",
            "nullable": true
          },
          "doc_id": {
            "type": "string"
          },
          "quality_score": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "source_ref": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "verified": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
...
      "MemoQueryRequest": {
        "type": "object",
        "required": [
          "session_token",
          "intent"
        ],
        "properties": {
          "intent": {
            "$ref": "#/components/schemas/PublicSearchIntent"
          },
          "session_token": {
            "type": "string"
          }
        }
      },
      "MemoQueryResponseEnvelope": {
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QueryResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
...
      "PublicFilter": {
        "type": "object",
        "description": "Filter constraint accepted by the public server API.",
        "required": [
          "field",
          "op",
          "value"
        ],
        "properties": {
          "field": {
            "type": "string",
            "description": "Field name to filter on (server allows `doc_id` and `chunk_id`)."
          },
          "op": {
            "$ref": "#/components/schemas/PublicFilterOp"
          },
          "value": {
            "type": "string",
            "description": "String-encoded filter value."
          }
        }
      },
      "PublicFilterOp": {
        "type": "string",
        "description": "Filter operator documented for the public server API.",
        "enum": [
          "eq",
          "contains"
        ]
      },
      "PublicSearchIntent": {
        "type": "object",
        "description": "Search intent describing how retrieval should run.",
        "required": [
          "retrieval_mode",
          "top_k",
          "hybrid_alpha",
          "filters"
        ],
        "properties": {
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicFilter"
            },
            "description": "Optional filter list (server allows `doc_id`/`chunk_id` with `eq`/`contains`)."
          },
          "hybrid_alpha": {
            "type": "number",
            "format": "float",
            "description": "Hybrid weight for vector vs. keyword scoring."
          },
          "indexing_config_id": {
            "type": "string",
            "description": "Optional indexing configuration selection (defaults to the project's default config).",
            "nullable": true
          },
          "query_embedding": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "Optional query embedding (required for `vector` and `hybrid` retrieval).",
            "nullable": true
          },
          "query_text": {
            "type": "string",
            "description": "Optional user query text (required for `fts` and `hybrid` retrieval).",
            "nullable": true
          },
          "reranker": {
            "allOf": [
              {
                "$ref": "#/components/schemas/coco_protocol.RerankerConfig"
              }
            ],
            "nullable": true
          },
          "retrieval_mode": {
            "$ref": "#/components/schemas/coco_protocol.RetrievalMode"
          },
          "top_k": {
            "type": "integer",
            "format": "int32",
            "description": "Number of candidates to return.",
            "minimum": 0
          }
        }
      },
      "QueryRequest": {
        "type": "object",
        "required": [
          "intent"
        ],
        "properties": {
          "indexing_config": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IndexingConfig"
              }
            ],
            "nullable": true
          },
          "indexing_config_id": {
            "type": "string",
            "nullable": true
          },
          "intent": {
            "$ref": "#/components/schemas/PublicSearchIntent"
          },
          "retrieval_config": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RetrievalConfig"
              }
            ],
            "nullable": true
          }
        }
      },
      "QueryResponse": {
        "type": "object",
        "required": [
          "results"
        ],
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            }
          }
        }
      },
      "QueryResponseEnvelope": {
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QueryResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
...
      "ResponseMeta": {
        "type": "object",
        "description": "Response metadata for request-level status.",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ResponseStatus"
          }
        }
      },
      "ResponseStatus": {
        "type": "string",
        "description": "Response freshness indicator.",
        "enum": [
          "fresh",
          "stale"
        ]
      },
...
      "SearchHit": {
        "type": "object",
        "description": "A search hit with metadata and chunk payload.",
        "required": [
          "meta",
          "chunk"
        ],
        "properties": {
          "chunk": {
            "$ref": "#/components/schemas/Chunk"
          },
          "meta": {
            "$ref": "#/components/schemas/SearchHitMeta"
          }
        }
      },
      "SearchHitMeta": {
        "type": "object",
        "description": "Metadata associated with a search hit.",
        "required": [
          "score"
        ],
        "properties": {
          "quality": {
            "type": "number",
            "format": "float",
            "description": "Optional quality indicator.",
            "nullable": true
          },
          "score": {
            "type": "number",
            "format": "float",
            "description": "Similarity or relevance score."
          },
          "verified": {
            "type": "boolean",
            "description": "Whether the result has been verified.",
            "nullable": true
          }
        }
      },
      "TextSpan": {
        "type": "object",
        "description": "A span inside a document, expressed as byte offsets.",
        "required": [
          "start",
          "end"
        ],
        "properties": {
          "end": {
            "type": "integer",
            "description": "Exclusive end byte offset.",
            "minimum": 0
          },
          "start": {
            "type": "integer",
            "description": "Inclusive start byte offset.",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}